jenkins

+++ Jenkinsfile
@@ -87,10 +87,14 @@ pipeline {
             steps {
                 script {
                     sh "docker rmi ${dockerRepoName}:${majorVersion}.$BUILD_NUMBER"
-                    sh "docker rmi ${dockerRegistryNoProto}/${dockerRepoName}:${majorVersion}.$BUILD_NUMBER"
                     sh "docker rmi ${dockerRepoName}:latest"
-                    sh "docker rmi ${dockerRegistryNoProto}/${dockerRepoName}:latest"
-                    sh "y | docker image prune"
+
+                    if (params.deployEnv == "${stagingEnv}") {
+                        sh "docker rmi ${dockerRegistryNoProto}/${dockerRepoName}:${majorVersion}.$BUILD_NUMBER"
+                        sh "docker rmi ${dockerRegistryNoProto}/${dockerRepoName}:latest"
+                    }
+
+                    sh "yes | docker image prune"
                 }
             }
         }
